Restructure `TestCaseImpactItem` table
This table currently contains near 1 billion rows and consumes 90% of the database size. Each index of this table occupies a few gigabytes of disk space.
With previous structure, indexes of the table use two times more disk space than the table's data.
The Id column has no use for this table. We need a useful clustered index to reduce the key lookups of the queries.
Also the `Id` column was of type `int`, and we were running out of ids.

I solved the problem by removing the `Id` column and creating a composite primary key.
Two big indexes removed, and the new PK is used instread of them.