Reset DbContext before every retry
Currently there is no way to reset the DbContext itself (EF Core 5 supports this).
The best way is to create a new instance of DbContext. Since we cannot
replace a created instance in the DI container (service provider), I made
DbContext *transient*, and it's container (BaseDbService) *scoped*, so when
we need to recreate the DbContext, we request a new instance from ServiceProvder.