Introduce a new table, "ProductLine", to identify the test branch
The role of `AzureProductBuildDefinitionId`, replaced with `ProductLine`.
For every value of `AzureProductBuildDefinitionId`, a new row inserted in
`ProductLine` table. To minimize changes to database, same values of
`AzureProductBuildDefinitionId` is used as `ProductLine.Id` (rebuilding keys
of the "TestCaseImpactItem" tables would takes hours to complete), and
by default, name of the product build is used as the `ProductLine.Key`